+++ lib/src/core/extensions/text_extension.dart
@@ -1,72 +1,73 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 
+///Text extension to directly style on Text
 extension TextExtension on Text {
   ///The size of glyphs (in logical pixels) to use when painting the text.
   ///
   ///Sets [fontSize] to 96 logical pixel and [fontWeight] to light
-  Text get h1 =>
-      _copyWith(style: TextStyle(fontSize: 96.0, fontWeight: FontWeight.w300));
+  Text get h1 => _copyWith(
+      style: const TextStyle(fontSize: 96.0, fontWeight: FontWeight.w300));
 
   ///The size of glyphs (in logical pixels) to use when painting the text.
   ///
   ///Sets [fontSize] to 60 logical pixel and [fontWeight] to light
-  Text get h2 =>
-      _copyWith(style: TextStyle(fontSize: 60.0, fontWeight: FontWeight.w300));
+  Text get h2 => _copyWith(
+      style: const TextStyle(fontSize: 60.0, fontWeight: FontWeight.w300));
 
   ///The size of glyphs (in logical pixels) to use when painting the text.
   ///
   ///Sets [fontSize] to 48 logical pixel and [fontWeight] to regular
-  Text get h3 => _copyWith(style: TextStyle(fontSize: 48.0));
+  Text get h3 => _copyWith(style: const TextStyle(fontSize: 48.0));
 
   ///The size of glyphs (in logical pixels) to use when painting the text.
   ///
   ///Sets [fontSize] to 34 logical pixel and [fontWeight] to regular
-  Text get h4 => _copyWith(style: TextStyle(fontSize: 34.0));
+  Text get h4 => _copyWith(style: const TextStyle(fontSize: 34.0));
 
   ///The size of glyphs (in logical pixels) to use when painting the text.
   ///
   ///Sets [fontSize] to 24 logical pixel and [fontWeight] to regular
-  Text get h5 => _copyWith(style: TextStyle(fontSize: 24.0));
+  Text get h5 => _copyWith(style: const TextStyle(fontSize: 24.0));
 
   ///The size of glyphs (in logical pixels) to use when painting the text.
   ///
   ///Sets [fontSize] to 20 logical pixel and [fontWeight] to medium
-  Text get h6 =>
-      _copyWith(style: TextStyle(fontSize: 20.0, fontWeight: FontWeight.w500));
+  Text get h6 => _copyWith(
+      style: const TextStyle(fontSize: 20.0, fontWeight: FontWeight.w500));
 
   ///The size of glyphs (in logical pixels) to use when painting the text.
   ///
   ///Sets [fontSize] to 16 logical pixel and [fontWeight] to Light
-  Text get body1 => _copyWith(style: TextStyle(fontSize: 16.0));
+  Text get body1 => _copyWith(style: const TextStyle(fontSize: 16.0));
 
   ///The size of glyphs (in logical pixels) to use when painting the text.
   ///
   ///Sets [fontSize] to 14 logical pixel and [fontWeight] to Light
-  Text get body2 =>
-      _copyWith(style: TextStyle(fontSize: 14.0, fontWeight: FontWeight.w500));
+  Text get body2 => _copyWith(
+      style: const TextStyle(fontSize: 14.0, fontWeight: FontWeight.w500));
 
   ///The size of glyphs (in logical pixels) to use when painting the text.
   ///
   ///Sets [fontSize] to 12 logical pixel and [fontWeight] to Light
-  Text get caption1 => _copyWith(style: TextStyle(fontSize: 12.0));
+  Text get caption1 => _copyWith(style: const TextStyle(fontSize: 12.0));
 
   ///The size of glyphs (in logical pixels) to use when painting the text.
   ///
   ///Sets [fontSize] to 10 logical pixel and [fontWeight] to Light
-  Text get caption2 => _copyWith(style: TextStyle(fontSize: 10.0));
+  Text get caption2 => _copyWith(style: const TextStyle(fontSize: 10.0));
 
   /// Sets [lineHeight] to 0.75
-  Text get heightTight => _copyWith(style: TextStyle(height: 0.75));
+  Text get heightTight => _copyWith(style: const TextStyle(height: 0.75));
 
   /// Sets [lineHeight] to 0.875
-  Text get heightSnug => _copyWith(style: TextStyle(height: 0.875));
+  Text get heightSnug => _copyWith(style: const TextStyle(height: 0.875));
 
   /// Sets [lineHeight] to 1.25
-  Text get heightRelaxed => _copyWith(style: TextStyle(height: 1.25));
+  Text get heightRelaxed => _copyWith(style: const TextStyle(height: 1.25));
 
   /// Sets [lineHeight] to 1.5
-  Text get heightLoose => _copyWith(style: TextStyle(height: 1.5));
+  Text get heightLoose => _copyWith(style: const TextStyle(height: 1.5));
 
   /// Sets custom [lineHeight] with [height]
   Text lineHeight(double height) => _copyWith(style: TextStyle(height: height));
@@ -117,11 +118,6 @@ extension TextExtension on Text {
   ///The text flows from right to left (e.g. Arabic, Hebrew).
   Text get rtl => _copyWith(textDirection: TextDirection.rtl);
 
-  ///The directionality of the text.
-  ///
-  ///The text flows from left to right (e.g., English, French).
-  Text get ltr => _copyWith(textDirection: TextDirection.ltr);
-
   ///Whether the text should break at soft line breaks.
   ///
   ///If false, the glyphs in the text will be positioned as if there was unlimited horizontal space.
@@ -151,50 +147,51 @@ extension TextExtension on Text {
   ///
   /// Draw a line underneath each line of text
   Text get underline =>
-      _copyWith(style: TextStyle(decoration: TextDecoration.underline));
+      _copyWith(style: const TextStyle(decoration: TextDecoration.underline));
 
   /// The decorations to paint near the text
   ///
   /// Draw a line through each line of text
   Text get lineThrough =>
-      _copyWith(style: TextStyle(decoration: TextDecoration.lineThrough));
+      _copyWith(style: const TextStyle(decoration: TextDecoration.lineThrough));
 
   /// The decorations to paint near the text
   ///
   /// Draw a line above each line of text
   Text get overline =>
-      _copyWith(style: TextStyle(decoration: TextDecoration.overline));
+      _copyWith(style: const TextStyle(decoration: TextDecoration.overline));
 
   /// The decorations to paint near the text
   ///
   ///Draw a [dashed] line underneath each line of text
-  Text get dashed =>
-      _copyWith(style: TextStyle(decorationStyle: TextDecorationStyle.dashed));
+  Text get dashed => _copyWith(
+      style: const TextStyle(decorationStyle: TextDecorationStyle.dashed));
 
   /// The decorations to paint near the text
   ///
   ///Draw a [doted] line underneath each line of text
-  Text get dotted =>
-      _copyWith(style: TextStyle(decorationStyle: TextDecorationStyle.dotted));
+  Text get dotted => _copyWith(
+      style: const TextStyle(decorationStyle: TextDecorationStyle.dotted));
 
   /// The decorations to paint near the text
   ///
   ///Draw a [double] means two line underneath each line of text
-  Text get doubleLine =>
-      _copyWith(style: TextStyle(decorationStyle: TextDecorationStyle.double));
+  Text get doubleLine => _copyWith(
+      style: const TextStyle(decorationStyle: TextDecorationStyle.double));
 
   /// The decorations to paint near the text
   ///
   ///Draw a [soild] line underneath each line of text
-  Text get solid =>
-      _copyWith(style: TextStyle(decorationStyle: TextDecorationStyle.solid));
+  Text get solid => _copyWith(
+      style: const TextStyle(decorationStyle: TextDecorationStyle.solid));
 
   /// The decorations to paint near the text
   ///
   ///Draw a [wavy] means sinusoidal line underneath each line of text
-  Text get wavy =>
-      _copyWith(style: TextStyle(decorationStyle: TextDecorationStyle.wavy));
+  Text get wavy => _copyWith(
+      style: const TextStyle(decorationStyle: TextDecorationStyle.wavy));
 
+  ///Sets the color in which to paint the text decorations.
   Text textDecoraionColor({Color decorationColor}) =>
       _copyWith(style: TextStyle(decorationColor: decorationColor));
 
@@ -202,7 +199,7 @@ extension TextExtension on Text {
   Text fontFamily({String fontFamily}) =>
       _copyWith(style: TextStyle(fontFamily: fontFamily));
 
-  //set [textScaleFactor] to text
+  ///set [textScaleFactor] to text
   Text textScaleFactor({double textScaleFactor}) =>
       _copyWith(textScaleFactor: textScaleFactor);
 
@@ -243,22 +240,26 @@ extension TextExtension on Text {
       _copyWith(style: TextStyle(letterSpacing: spacing));
 
   /// Sets [letterSpacing] to -3.0
-  Text get tightestLetter => _copyWith(style: TextStyle(letterSpacing: -3.0));
+  Text get tightestLetter =>
+      _copyWith(style: const TextStyle(letterSpacing: -3.0));
 
   /// Sets [letterSpacing] to -2.0
-  Text get tighterLetter => _copyWith(style: TextStyle(letterSpacing: -2.0));
+  Text get tighterLetter =>
+      _copyWith(style: const TextStyle(letterSpacing: -2.0));
 
   /// Sets [letterSpacing] to -1.0
-  Text get tightLetter => _copyWith(style: TextStyle(letterSpacing: -1.0));
+  Text get tightLetter =>
+      _copyWith(style: const TextStyle(letterSpacing: -1.0));
 
   /// Sets [letterSpacing] to 1.0
-  Text get wideLetter => _copyWith(style: TextStyle(letterSpacing: 1.0));
+  Text get wideLetter => _copyWith(style: const TextStyle(letterSpacing: 1.0));
 
   /// Sets [letterSpacing] to 2.0
-  Text get widerLetter => _copyWith(style: TextStyle(letterSpacing: 2.0));
+  Text get widerLetter => _copyWith(style: const TextStyle(letterSpacing: 2.0));
 
   /// Sets [letterSpacing] to 3.0
-  Text get widestLetter => _copyWith(style: TextStyle(letterSpacing: 3.0));
+  Text get widestLetter =>
+      _copyWith(style: const TextStyle(letterSpacing: 3.0));
 
   ///The amount of space (in logical pixels) to add at each sequence of white-space
   ///
@@ -268,52 +269,58 @@ extension TextExtension on Text {
       _copyWith(style: TextStyle(wordSpacing: spacing));
 
   /// Sets [WordSpacing] to -3.0
-  Text get tightestWord => _copyWith(style: TextStyle(wordSpacing: -3.0));
+  Text get tightestWord => _copyWith(style: const TextStyle(wordSpacing: -3.0));
 
   /// Sets [WordSpacing] to -2.0
-  Text get tighterWord => _copyWith(style: TextStyle(wordSpacing: -2.0));
+  Text get tighterWord => _copyWith(style: const TextStyle(wordSpacing: -2.0));
 
   /// Sets [WordSpacing] to -1.0
-  Text get tightWord => _copyWith(style: TextStyle(wordSpacing: -1.0));
+  Text get tightWord => _copyWith(style: const TextStyle(wordSpacing: -1.0));
 
   /// Sets [WordSpacing] to 1.0
-  Text get wideWord => _copyWith(style: TextStyle(wordSpacing: 1.0));
+  Text get wideWord => _copyWith(style: const TextStyle(wordSpacing: 1.0));
 
   /// Sets [WordSpacing] to 2.0
-  Text get widerWord => _copyWith(style: TextStyle(wordSpacing: 2.0));
+  Text get widerWord => _copyWith(style: const TextStyle(wordSpacing: 2.0));
 
   /// Sets [WordSpacing] to 3.0
-  Text get widestWord => _copyWith(style: TextStyle(wordSpacing: 3.0));
+  Text get widestWord => _copyWith(style: const TextStyle(wordSpacing: 3.0));
 
   ///Sets [fontWeight] to [FontWeight.w100]
-  Text get thin => _copyWith(style: TextStyle(fontWeight: FontWeight.w100));
+  Text get thin =>
+      _copyWith(style: const TextStyle(fontWeight: FontWeight.w100));
 
   ///Sets [fontWeight] to [FontWeight.w200]
   Text get extraLight =>
-      _copyWith(style: TextStyle(fontWeight: FontWeight.w200));
+      _copyWith(style: const TextStyle(fontWeight: FontWeight.w200));
 
   ///Sets [fontWeight] to [FontWeight.w300]
-  Text get light => _copyWith(style: TextStyle(fontWeight: FontWeight.w300));
+  Text get light =>
+      _copyWith(style: const TextStyle(fontWeight: FontWeight.w300));
 
   ///Sets [fontWeight] to [FontWeight.w400]
-  Text get normal => _copyWith(style: TextStyle(fontWeight: FontWeight.w400));
+  Text get normal =>
+      _copyWith(style: const TextStyle(fontWeight: FontWeight.w400));
 
   ///Sets [fontWeight] to [FontWeight.w500]
-  Text get medium => _copyWith(style: TextStyle(fontWeight: FontWeight.w500));
+  Text get medium =>
+      _copyWith(style: const TextStyle(fontWeight: FontWeight.w500));
 
   ///Sets [fontWeight] to [FontWeight.w600]
-  Text get bold => _copyWith(style: TextStyle(fontWeight: FontWeight.w600));
+  Text get bold =>
+      _copyWith(style: const TextStyle(fontWeight: FontWeight.w600));
 
   ///Sets [fontWeight] to [FontWeight.w700]
-  Text get semiBold => _copyWith(style: TextStyle(fontWeight: FontWeight.w700));
+  Text get semiBold =>
+      _copyWith(style: const TextStyle(fontWeight: FontWeight.w700));
 
   ///Sets [fontWeight] to [FontWeight.w800]
   Text get extraBold =>
-      _copyWith(style: TextStyle(fontWeight: FontWeight.w800));
+      _copyWith(style: const TextStyle(fontWeight: FontWeight.w800));
 
   ///Sets [fontWeight] to [FontWeight.w900]
   Text get blackBold =>
-      _copyWith(style: TextStyle(fontWeight: FontWeight.w900));
+      _copyWith(style: const TextStyle(fontWeight: FontWeight.w900));
 
   Text _copyWith(
       {Key key,
@@ -328,7 +335,7 @@ extension TextExtension on Text {
       String semanticsLabel,
       TextWidthBasis textWidthBasis,
       TextStyle style}) {
-    return Text(this.data,
+    return Text(data,
         key: key ?? this.key,
         strutStyle: strutStyle ?? this.strutStyle,
         textAlign: textAlign ?? this.textAlign,