+++ lib/src/core/extensions/widget_extension.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 
+///Widget extension to entend widget functionality
 extension WidgetExtensions on Widget {
   ///Wrap up any widget without using [center] widget
   ///
@@ -19,18 +20,28 @@ extension WidgetExtensions on Widget {
   Opacity opcity({@required double opacity}) =>
       Opacity(opacity: opacity, child: this);
 
+  ///
   ///Get 0% [Opacity] which means fully invisible
+  ///
   Opacity get opacity0 => Opacity(opacity: 0.0, child: this);
 
+  ///
   ///Get 25% [Opacity] which means paritally visible
+  ///
   Opacity get opacity25 => Opacity(opacity: 0.25, child: this);
 
+  ///
   ///Get 50% [Opacity] which means half visible
+  ///
   Opacity get opacity50 => Opacity(opacity: 0.50, child: this);
 
+  ///
   ///Get 75% [Opacity] which means paritally invisible
+  ///
   Opacity get opacity75 => Opacity(opacity: 0.75, child: this);
 
+  ///
   ///Get 100% [Opacity] which means fully visible
+  ///
   Opacity get opacity100 => Opacity(opacity: 1.0, child: this);
 }