+++ lib/src/core/mixins/pixel_mixin.dart
@@ -1,15 +1,41 @@
+///Predefined logical pixel mixin
 mixin FxPixelMixin {
+  ///Gives logical pixel 0
   static const double dp0 = 0;
+
+  ///Gives logical pixel 4
   static const double dp4 = 4;
+
+  ///Gives logical pixel 8
   static const double dp8 = 8;
+
+  ///Gives logical pixel 12
   static const double dp12 = 12;
+
+  ///Gives logical pixel 16
   static const double dp16 = 16;
+
+  ///Gives logical pixel 20
   static const double dp20 = 20;
+
+  ///Gives logical pixel 24
   static const double dp24 = 24;
+
+  ///Gives logical pixel 32
   static const double dp32 = 32;
+
+  ///Gives logical pixel 48
   static const double dp48 = 48;
+
+  ///Gives logical pixel 64
   static const double dp64 = 64;
+
+  ///Gives logical pixel 11
   static const double dp112 = 112;
+
+  ///Gives logical pixel 14
   static const double dp144 = 144;
+
+  ///Gives logical pixel 17
   static const double dp176 = 176;
 }