+++ lib/src/fx_extensions.dart
@@ -1,5 +1,6 @@
-export 'core/extensions/padding_extension.dart';
-export 'core/extensions/widget_extension.dart';
-export 'core/extensions/String_extension.dart';
-export 'core/extensions/Text_extension.dart';
 export 'core/extensions/context_extension.dart';
+export 'core/extensions/padding_extension.dart';
+export 'core/extensions/string_extension.dart';
+export 'core/extensions/text_extension.dart';
+export 'core/extensions/widget_extension.dart';
+export 'core/extensions/container_extension.dart';