+++ lib/src/fx_widgets.dart
@@ -1,1 +1,1 @@
-export '../src/widgets/button/fx_button.dart';
+