+++ CHANGELOG.md
@@ -1,2 +1,14 @@
 ## [0.0.1] - Initial release
 ## [0.0.2] - String and Text extension methods added
+## [0.0.3] - March 22,2021
+    - Added BuildContext extensions
+        * Get MediaQuery size
+        * Get Orientation
+    * Added Widget extensions
+        * Opcaity [opcaity25],[opcaity50],[opcaity75],[opcaity100]
+    - Decoration
+    - Box Shadow
+## [0.0.4] - March 23,2021
+    - Container extensions
+    - BorderRadius
+    ** Bug fixed